switch to fp16 as the go-to float dtype (i.e. how to infer a python float literal, or how to convert ints to floats). to try and simplify the conversion process (which currently starts by reading an fp16 torchscript scriptmodule, converts to fp32 CoreML, then optimizes back to fp16 CoreML). conversion completes, but XCode complains that the .mlpackage is missing inputSchema, and when you try to load the model in python you get complaints due to the changes made to type domains on upsample function

+++ coremltools/converters/mil/mil/ops/defs/iOS15/elementwise_binary.py
@@ -25,7 +25,7 @@ class elementwise_binary(Operation):
     )
 
     type_domains = {
-        "T": (types.fp16, types.fp32, types.int32),
+        "T": (types.fp16, types.fp16, types.int32),
     }
 
     def type_inference(self):
